requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
alembic==1.16.4
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
APScheduler==3.11.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
audio-recorder-streamlit==0.0.10
audioread==3.0.1
av==14.4.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bertopic==0.15.0
bidict==0.23.1
blinker==1.9.0
blis==0.7.11
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.1.8
cloudpathlib==0.21.1
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dataclasses-json==0.6.7
decorator==5.2.1
deep-sort-realtime==1.3.2
deep-translator==1.11.4
deepface==0.0.89
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docx2txt==0.9
durationpy==0.10
easyocr==1.7.2
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
facenet-pytorch==2.6.0
fastapi==0.116.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.19.1
fire==0.7.0
firebase-admin==6.9.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-PyMongo==3.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.58.5
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gast==0.4.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
google-cloud-core==2.4.3
google-cloud-documentai==3.7.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
graphviz==0.21
greenlet==3.2.4
grpcio==1.73.1
grpcio-status==1.73.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
h5py==3.14.0
hdbscan==0.8.29
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==5.2.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
importlib-metadata==6.11.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.30
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.31
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
lxml==6.0.2
lz4==4.4.4
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==5.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
mtcnn==0.1.1
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.0.1
networkx==3.4.2
ninja==1.13.0
nltk==3.8.1
numba==0.60.0
numpy==1.24.4
oauthlib==3.3.1
onnxruntime==1.23.0
openai==2.15.0
openai-whisper==20250625
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pathlib_abc==0.1.1
pathy==0.11.0
pdfminer.six==20250506
pdfplumber==0.11.7
pickle-mixin==1.0.2
pillow==10.2.0
pipwin==0.5.2
platformdirs==4.4.0
playsound==1.3.0
plotly==6.2.0
pooch==1.8.2
postgrest==2.28.3
posthog==5.4.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.13.3
pydantic-settings==2.11.0
pydantic_core==2.46.3
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pyiceberg==0.11.1
pyjsparser==2.7.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pymongo==4.13.2
Pympler==1.1
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyppeteer==2.0.0
PyPrind==2.11.3
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
pyreadline3==3.5.4
pyroaring==1.0.4
pyrsistent==0.20.0
pyserial==3.5
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.13.0
realtime==2.28.3
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.1.1
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
simpleitk==2.5.3
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
srsly==2.5.1
srt==3.5.3
starlette==0.46.2
storage3==2.28.3
streamlit==1.48.1
streamlit-audiorec==0.1.3
streamlit-chat==0.1.1
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
streamlit-webrtc==0.63.3
streamlit_mic_recorder==0.0.8
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tesseract==0.1.3
tf-keras==2.15.0
thinc==8.1.12
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.22
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.9.1
torchaudio==2.8.0
torchvision==0.24.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.57.3
tweepy==4.16.0
typer==0.7.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==4.3.1
ultralytics==8.3.167
ultralytics-thop==2.0.14
umap-learn==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
vosk==0.3.45
wasabi==1.1.3
watchdog==4.0.2
watchfiles==1.1.0
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.14.1
wsproto==1.2.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
